Fix edit item type match

+++ Room2Room/Models/DTOs/EditItem.cs
@@ -7,7 +7,7 @@ namespace Room2Room.Models.DTOs
         public int Id { get; set; }
         public string ItemName { get; set; } = "";
         public string ItemDescription { get; set; } = "";
-        [Range(typeof(decimal), "0", "10000")]
+        [Range(typeof(double), "0", "10000")]
         public double Price { get; set; }
         public int CategoryId { get; set; }
         public string Status { get; set; } = "Available";
